refactor(Venue) naming, namespaces and repository contract
