+++ src/Application/Feature/Venue/Requests/VenueSectionsRequest.cs
@@ -1,0 +1,3 @@
+﻿namespace Ticketing.Application.Feature.Venue.Requests;
+
+public record VenueSectionsRequest(int VenueId);
