+++ src/Application/Feature/Venue/Requests/VenuesQueryRequest.cs
@@ -1,0 +1,3 @@
+﻿namespace Ticketing.Application.Feature.Venue.Requests;
+
+public record VenuesQueryRequest();
