+++ src/Application/Feature/Venue/Responses/VenueSeactionsResponse.cs
@@ -1,0 +1,5 @@
+﻿using Ticketing.Domain.Entities.Venue;
+
+namespace Ticketing.Application.Feature.Venue.Responses;
+
+public record VenueSectionsResponse(IEnumerable<Section> Sections);
