+++ src/Application/Feature/Venue/Responses/VenuesQueryResponse.cs
@@ -1,0 +1,3 @@
+﻿namespace Ticketing.Application.Feature.Venue.Responses;
+
+public record VenuesQueryResponse(IEnumerable<Domain.Entities.Venue.Venue> Venues);
