+++ src/Application/Feature/Venue/VenueSectionsQuery.cs
@@ -1,10 +1,10 @@
 ﻿using Ticketing.Application.CQRS;
+using Ticketing.Application.Feature.Venue.Requests;
 using Ticketing.Domain.Entities.Venue;
 using Ticketing.Domain.Interfaces.Repositories;
 
-namespace Ticketing.Application;
+namespace Ticketing.Application.Feature.Venue;
 
-public record VenueSectionsRequest(int VenueId);
 public record VenueSectionsResponse(IEnumerable<Section> Sections);
 
 /// <summary>
@@ -12,17 +12,17 @@ public record VenueSectionsResponse(IEnumerable<Section> Sections);
 /// </summary>
 public class VenueSectionsQuery : IQueryHandler<VenueSectionsRequest, VenueSectionsResponse>
 {
-    private readonly IVenueRepository repository;
+    private readonly IVenueRepository _venueRepository;
 
-    public VenueSectionsQuery(IVenueRepository repository)
+    public VenueSectionsQuery(IVenueRepository venueRepository)
     {
-        this.repository = repository;
+        this._venueRepository = venueRepository;
     }
 
     public async Task<VenueSectionsResponse> ExecuteAsync(VenueSectionsRequest request, CancellationToken cancellation)
     {
-        var venue = await this.repository.GetAsync(request.VenueId, cancellation);
+        var venue = await this._venueRepository.GetWithSectionsAsync(request.VenueId, cancellation);
 
-        return new VenueSectionsResponse(venue?.Sections?.ToList()!);
+        return new VenueSectionsResponse(venue?.Sections!);
     }
 }
