+++ src/Application/Feature/Venue/VenuesQuery.cs
@@ -1,25 +1,22 @@
 ﻿using Ticketing.Application.CQRS;
-using Ticketing.Domain.Entities.Venue;
-using Ticketing.Domain.Interfaces;
+using Ticketing.Application.Feature.Venue.Requests;
+using Ticketing.Application.Feature.Venue.Responses;
+using Ticketing.Domain.Interfaces.Repositories;
 
-namespace Ticketing.Application;
-
-public record VenuesQueryRequest();
-
-public record VenuesQueryResponse(IEnumerable<Venue> Venues);
+namespace Ticketing.Application.Feature.Venue;
 
 public class VenuesQuery : IQueryHandler<VenuesQueryRequest, VenuesQueryResponse>
 {
-    private readonly IRepository<Venue, int> repository;
+    private readonly IVenueRepository _venueRepository;
 
-    public VenuesQuery(IRepository<Venue, int> repository)
+    public VenuesQuery(IVenueRepository venueRepository)
     {
-        this.repository = repository;
+        this._venueRepository = venueRepository;
     }
 
     public async Task<VenuesQueryResponse> ExecuteAsync(VenuesQueryRequest request, CancellationToken cancellation)
     {
-        var result = await this.repository.ListAsync(cancellation);
+        var result = await this._venueRepository.ListAsync(cancellation);
 
         return new VenuesQueryResponse(result!);
     }
