+++ src/Domain/Interfaces/Repositories/IVenueRepository.cs
@@ -4,5 +4,5 @@ namespace Ticketing.Domain.Interfaces.Repositories;
 
 public interface IVenueRepository : IRepository<Venue, int>
 {
-
+    Task<Venue> GetWithSectionsAsync(int venueId, CancellationToken cancellation);
 }
